feat(logging): implement ELK stack integration and distributed tracing
- Add Elasticsearch, Logstash, and Kibana services to docker-compose.yml

- Create Logstash pipeline config for parsing incoming JSON logs

- Introduce HttpLoggingFilter and FeignLoggingInterceptor in common-lib

- Implement structured JSON logging via HttpLogEntity and Logstash encoder

- Add X-Correlation-Id propagation for end-to-end distributed tracing

- Refactor Feign interceptors (CorrelationId & AuthForwarding) for SRP

- Config logback-spring.xml to stream logs from all 5 microservices

- Remove legacy duplicate interceptors from payment and invoice services

+++ wallet-service/src/main/java/com/hk/hk/walletservice/dto/response/WalletResponse.java
@@ -2,14 +2,13 @@ package com.hk-fintech.hk.walletservice.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
-
 import java.math.BigDecimal;
 
 @Data
 @Builder
 public class WalletResponse {
     private String id;
-    private Integer userId;
+    private Long userId;
     private BigDecimal balance;
     private String currency;
 }
